Remove buddy allocation file
