+++ src/storage/CMakeLists.txt
@@ -10,7 +10,7 @@ add_subdirectory(wal)
 
 add_library(kuzu_storage
         OBJECT
-        free_space_manager_sorted.cpp
+        free_space_manager.cpp
         block_manager.cpp
         db_file_id.cpp
         file_handle.cpp
